+++ src/tk115/CMakeLists.txt
@@ -1,7 +1,5 @@
 SET(SOURCE_FILE
     main.c
-    mqtt.c
-    mqtt.h
     server_tk115.c
     server_tk115.h
     msg_sch_tk115.c
