+++ src/tk115/main.c
@@ -12,6 +12,7 @@
 #include "db.h"
 #include "session.h"
 #include "mqtt.h"
+#include "msg_proc_app.h"
 
 
 struct event_base *base = NULL;
@@ -79,7 +80,7 @@ int main(int argc, char **argv)
     	return -1;
     }
 
-    mqtt_initial();
+    mqtt_initial(app_handleApp2devMsg);
 
     rc = yunba_connect();
     if (rc)
