+++ src/tk115/msg_proc_app.c
@@ -178,95 +178,3 @@ int app_handleApp2devMsg(const char* topic, const char* data, const int len, voi
 	return 0;
 }
 
-void app_message_callback(struct mosquitto *mosq __attribute__((unused)), void *userdata, const struct mosquitto_message *message)
-{
-	if(message->payloadlen){
-		LOG_DEBUG("%s %p", message->topic, message->payload);
-	}else{
-		LOG_DEBUG("%s no payload(null)", message->topic);
-	}
-
-	LOG_INFO("recieve PUBLISH: %s", message->topic);
-
-	if(strncmp(message->topic,"app2dev/",strlen("app2dev/")) == 0)
-	{
-		app_handleApp2devMsg(message->topic, message->payload, message->payloadlen, userdata);
-	}
-	else
-	{
-		LOG_ERROR("Receive unknown PUBLISH");
-	}
-
-}
-
-void app_connect_callback(struct mosquitto *mosq, void *userdata, int rc)
-{
-	if(!rc)
-	{
-		LOG_INFO("Connect to MQTT server successfully");
-	}
-	else
-	{
-		//TODO: check whether mosquitto_connack_string here is OK
-		LOG_ERROR("Connect failed: result = %s", mosquitto_connack_string(rc));
-	}
-}
-
-void app_disconnect_callback(struct mosquitto *mosq __attribute__((unused)), void *userdata, int rc)
-{
-
-	if(rc)
-	{
-		LOG_ERROR("disconnect rc = %d(%s)\n",  rc, mosquitto_strerror(rc));
-
-		//mosquitto_reconnect(mosq);
-	}
-	else
-	{
-		LOG_INFO("client disconnect successfully");
-	}
-}
-
-
-void app_subscribe_callback(struct mosquitto *mosq __attribute__((unused)), void *userdata __attribute__((unused)), int mid, int qos_count, const int *granted_qos)
-{
-	LOG_DEBUG("Subscribed (mid: %d): %d", mid, granted_qos[0]);
-	for(int i=1; i<qos_count; i++){
-		LOG_DEBUG(", %d", granted_qos[i]);
-	}
-}
-
-void app_log_callback(struct mosquitto *mosq __attribute__((unused)), void *userdata __attribute__((unused)), int level, const char *str)
-{
-	switch (level)
-	{
-	case MOSQ_LOG_DEBUG:
-		LOG_DEBUG("%s", str);
-
-		break;
-	case MOSQ_LOG_INFO:
-	case MOSQ_LOG_NOTICE:
-	case MOSQ_LOG_SUBSCRIBE:
-	case MOSQ_LOG_UNSUBSCRIBE:
-		LOG_INFO("%s", str);
-		break;
-
-	case MOSQ_LOG_WARNING:
-		LOG_WARN("%s", str);
-		break;
-
-	case MOSQ_LOG_ERR:
-		LOG_ERROR("%s", str);
-		break;
-
-	default:
-		LOG_ERROR("unknown level log:%s", str);
-	}
-
-}
-
-void app_publish_callback(struct mosquitto *mosq __attribute__((unused)), void *userdata __attribute__((unused)), int mid __attribute__((unused)))
-{
-	LOG_DEBUG("Published mid: %d", mid);
-	LOG_INFO("Publish mid: %d successfully", mid);
-}
